+++ src/basis/type/DEFAULT_VALUE.js
@@ -1,0 +1,1 @@
+module.exports = {};
